Adding a new generic, re-usable conversion-gen example
This patch is taking conversion-gen out of the code-generator repo, and instead adding it here as a re-usable, generic example; for the sake of being able to re-use the conversion logic in other generators - this will be re-used immediately in [csi-proxy](https://github.com/kubernetes-csi/csi-proxy/tree/15613db6b0fa6b2b183e1ecc52b3bfd64b4a5005/client/api#detailed-breakdown-of-generated-files).

A new package, `k8s.io/gengo/examples/conversion-gen/generators/generator`, exposes a new `ConversionGenerator` public struct that fulfills the `Generator` interface, and can be re-used in other repos. That structure is wrapped in a private `genConversion` struct that adds the logic specific to kubernetes.

This patch could be adopted as-in in k8s itself, as a drop-in replacement for the current conversion-gen code, as demonstrated by.

Added a few tests.

Signed-off-by: Jean Rouge <rougej+github@gmail.com>